+++ changelog.md
@@ -1,3 +1,7 @@
-v0.5.2:
-- binary, unary ops codegen
-- funcs codegen
+v0.5.3:
+- variable reassignment
+- fixes
+- unary ops
+- more fixes
+- example `1_variables.fcy` now works!
+- added automated tests via `cargo test`